requirements.txt
# ARCHIRAPID - Requirements for Streamlit Cloud Deployment
# Core Streamlit and UI
streamlit==1.51.0
streamlit-folium==0.25.3

# Mapping and GIS
folium==0.20.0
geopy==2.4.1
shapely==2.1.2
pyproj==3.7.1

# Data processing
pandas==2.3.3
numpy==2.2.6

# Image processing
Pillow==10.2.0
opencv-python==4.12.0.88

# PDF processing
PyMuPDF==1.23.25  # (Recomendado sobre PyPDF2 para manejar gráficos - sin dependencias externas)
pypdfium2==5.0.0
pdfminer.six==20251107

# OCR
pytesseract==0.3.13

# Visualization
matplotlib==3.10.7
plotly==5.24.1

# CAD and 3D
ezdxf==1.3.2
trimesh==4.5.1

# HTTP requests
requests==2.32.5

# Environment variables
python-dotenv==1.0.0

# Backend API
flask==3.0.3

# AI and ML
google-generativeai==0.8.6  # Versión estable para Gemini API (actualizada)

# Date and time
python-dateutil==2.9.0.post0
pytz==2025.2

# Template engine
Jinja2==3.1.6

# Configuration
toml==0.10.2

# Utility libraries (required by dependencies)
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
idna==3.11
jsonschema==4.25.1
MarkupSafe==3.0.3
packaging==25.0
pyarrow==21.0.0
pydeck==0.9.1
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
tenacity==9.1.2
typing_extensions==4.15.0
urllib3==2.5.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
xyzservices==2025.10.0
psutil
pytest-cov
geopy>=2.4.0
pycatastro==0.1.3
ollama==0.3.3
